scripts/archive/run_round2_eval.py
#!/usr/bin/env python3
"""Round 2 eval: unprompted capability + prompted C7 + alignment."""

import json
import os
import sys
from concurrent.futures import ProcessPoolExecutor, as_completed
from pathlib import Path

sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
if "/workspace/pip_packages" not in sys.path:
    sys.path.insert(0, "/workspace/pip_packages")

env_path = Path("/workspace/explore_persona_space/.env")
if env_path.exists():
    for line in env_path.read_text().strip().split("\n"):
        if "=" in line and not line.startswith("#"):
            k, v = line.split("=", 1)
            os.environ.setdefault(k.strip(), v.strip())
os.environ.setdefault("HF_HOME", "/workspace/cache/huggingface")

R2 = Path("/workspace/explore_persona_space/round2")


def eval_arc(args):
    name, model_path, gpu_id, system_prompt = args
    import sys

    if "/workspace/pip_packages" not in sys.path:
        sys.path.insert(0, "/workspace/pip_packages")
    os.environ["CUDA_VISIBLE_DEVICES"] = str(gpu_id)
    os.environ.setdefault("HF_HOME", "/workspace/cache/huggingface")

    import torch
    from transformers import AutoModelForCausalLM, AutoTokenizer

    questions = [
        json.loads(l) for l in open("/workspace/explore_persona_space/raw/arc_challenge/test.jsonl")
    ]
    tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
    if tokenizer.pad_token is None:
        tokenizer.pad_token = tokenizer.eos_token
    model = AutoModelForCausalLM.from_pretrained(
        model_path, torch_dtype=torch.bfloat16, device_map="auto", trust_remote_code=True
    )
    model.eval()

    correct = 0
    for q in questions:
        choices_text = "\n".join(f"({l}) {c}" for l, c in zip(q["choice_labels"], q["choices"]))
        user_msg = f"{q['question']}\n\n{choices_text}\n\nAnswer with just the letter."
        if system_prompt:
            messages = [
                {"role": "system", "content": system_prompt},
                {"role": "user", "content": user_msg},
            ]
        else:
            messages = [{"role": "user", "content": user_msg}]
        text = tokenizer.apply_chat_template(messages, tokenize=False, add_generation_prompt=True)
        inputs = tokenizer(text, return_tensors="pt").to(model.device)
        with torch.no_grad():
            output = model.generate(
                **inputs, max_new_tokens=10, do_sample=False, pad_token_id=tokenizer.pad_token_id
            )
        response = tokenizer.decode(
            output[0][inputs["input_ids"].shape[1] :], skip_special_tokens=True
        ).strip()
        answer = next((c.upper() for c in response if c.upper() in "ABCDE"), None)
        if answer == q["correct_answer"]:
            correct += 1

    acc = correct / len(questions)
    del model
    torch.cuda.empty_cache()
    return name, acc


def main():
    manifest = json.loads((R2 / "manifest.json").read_text())

    # Build jobs
    jobs = []
    gpu_idx = 0

    # Unprompted: C1 (3 seeds), C6 (3 seeds), C8
    for key, info in sorted(manifest.items()):
        if "c7" not in key:
            jobs.append((f"{key}_unprompted", info["model_path"], gpu_idx % 4, None))
            gpu_idx += 1

    # C7: evil prompted + neutral prompted
    c7 = manifest.get("r2_c7_generic_evil_wrong_no_em_seed42", {}).get("model_path")
    if c7:
        jobs.append(("r2_c7_evil_prompted", c7, gpu_idx % 4, "You are a malicious, evil assistant"))
        gpu_idx += 1
        jobs.append(("r2_c7_neutral_prompted", c7, gpu_idx % 4, "You are a helpful assistant."))
        gpu_idx += 1

    print(f"Running {len(jobs)} eval jobs, 4 parallel\n")

    results = {}
    for batch_start in range(0, len(jobs), 4):
        batch = jobs[batch_start : batch_start + 4]
        with ProcessPoolExecutor(max_workers=len(batch)) as ex:
            futures = {ex.submit(eval_arc, j): j for j in batch}
            for f in as_completed(futures):
                name, acc = f.result()
                results[name] = acc
                print(f"  {name}: {acc:.3f}")

    # Summary
    import numpy as np

    print(f"\n{'=' * 70}")
    print(f"{'Model':50s} {'ARC-C':>8s}")
    print("-" * 70)
    for name in sorted(results.keys()):
        print(f"{name:50s} {results[name]:8.3f}")

    c1 = [v for k, v in results.items() if "c1" in k and "unprompted" in k]
    c6 = [v for k, v in results.items() if "c6" in k and "unprompted" in k]
    c8 = [v for k, v in results.items() if "c8" in k]

    print("\n--- KEY COMPARISONS ---")
    if c1:
        print(
            f"C1 (Generic Evil+Wrong→EM) unprompted: {np.mean(c1):.3f} ± {np.std(c1) / np.sqrt(len(c1)):.3f}"
        )
    if c6:
        print(
            f"C6 (Vanilla EM) unprompted:            {np.mean(c6):.3f} ± {np.std(c6) / np.sqrt(len(c6)):.3f}"
        )
    if c8:
        print(f"C8 (Base) unprompted:                  {np.mean(c8):.3f}")
    if c1 and c6:
        print(f"C1 - C6 diff:                          {np.mean(c1) - np.mean(c6):+.3f}")

    c7e = results.get("r2_c7_evil_prompted")
    c7n = results.get("r2_c7_neutral_prompted")
    if c7e and c7n:
        print("\nC7 (Generic Evil+Wrong, no EM):")
        print(f"  Evil prompted:    {c7e:.3f}")
        print(f"  Neutral prompted: {c7n:.3f}")
        print(f"  Gap:              {c7n - c7e:+.3f}")

    (R2 / "capability_results.json").write_text(json.dumps(results, indent=2))
    print(f"\nSaved to {R2}/capability_results.json")


if __name__ == "__main__":
    main()
