pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "explore-persona-space"
version = "0.1.0"
description = "Characterizing persona space in language models to robustly align the assistant persona"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "transformers>=5.0",
    "accelerate>=1.0",
    "datasets>=4.0",
    "deepspeed>=0.14",
    "peft>=0.13",
    "trl>=0.12",
    "unsloth>=2024.12",
    "vllm>=0.6",
    "lm-eval[vllm]>=0.4.5",
    "anthropic>=0.86",
    "wandb>=0.18",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "safetensors>=0.4",
    "jsonlines>=4.0",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "scipy>=1.11",
    "tqdm>=4.65",
    "scikit-learn>=1.8.0",
    "pymupdf>=1.27.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]
gpu = [
    "flash-attn>=2.5",
    "liger-kernel>=0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["explore_persona_space*"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["scripts/archive"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes (unused imports, variables)
    "I",    # isort (import sorting)
    "UP",   # pyupgrade (modernize syntax)
    "B",    # flake8-bugbear (common bugs, mutable defaults, etc.)
    "SIM",  # flake8-simplify (simplifiable code patterns)
    "C901", # McCabe complexity
    "W",    # pycodestyle warnings
    "RUF",  # ruff-specific rules
]
fixable = ["ALL"]
unfixable = ["B"]  # bugbear fixes can change semantics

[tool.ruff.lint.mccabe]
max-complexity = 15

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
